worker/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
redis
boto3
yt-dlp
ffmpeg-python
